First request diaplay
The current update is teh first time the application has pulled data scuccessfully from the API. The API key and base URL from the configuration
is currently not being accessed by the application. This is currently being bypassed by having the two variables hard coded in the application. This is obviosly not the best practice and will be updated before the final deployment. The news sources from the API are at this stahe successfully rendereed on the index.html page